Updated to be python3  compatible

+++ utils/progress.py
@@ -9,8 +9,8 @@ class WorkSplitter(object):
 
     def section(self, name):
         name_length = len(name)
-        left_length = (self.columns-name_length)/2
-        right_length = self.columns-name_length-left_length
+        left_length = int((self.columns-name_length)/2)
+        right_length = int(self.columns-name_length-left_length)
 
         output = '='*self.columns+'\n' \
                  + "|"+' '*(left_length-1)+name+' '*(right_length-1)+'|\n'\
@@ -20,15 +20,15 @@ class WorkSplitter(object):
 
     def subsection(self, name):
         name_length = len(name)
-        left_length = (self.columns-name_length)/2
-        right_length = self.columns-name_length-left_length
+        left_length = int((self.columns-name_length)/2)
+        right_length = int(self.columns-name_length-left_length)
 
         output = '#' * (left_length-1) + ' ' + name + ' ' + '#' * (right_length-1) + '\n'
         print(output)
 
     def subsubsection(self, name):
         name_length = len(name)
-        left_length = (self.columns-name_length)/2
+        left_length = int((self.columns-name_length)/2)
         right_length = self.columns-name_length-left_length
 
         output = '-' * (left_length-1) + ' ' + name + ' ' + '-' * (right_length-1) + '\n'
